Move files to utils folder
